fixed the dashboard
