+++ src/App.css
@@ -4,6 +4,7 @@
     --fontFamily: 'Inter', sans-serif;
     --textDark: #1111111;
 <<<<<<< HEAD
+<<<<<<< HEAD
     --light: #F7F7F7;
     --gold: #B5B500;
     --red: #EE4D1D;
@@ -11,4 +12,9 @@
     --bgDark: #263238;
 
 >>>>>>> ea36324 (Made changes on the ui)
+=======
+    --light: #F7F7F7;
+    --gold: #B5B500;
+    --red: #EE4D1D;
+>>>>>>> 97c5236 (fixed the dashboard)
 }
