+++ src/components/newsidebar/newsidebar.css
@@ -7,7 +7,7 @@
 .sidebar {
     width: 300px;
     height: 100vh;
-    background-color: #263238;  
+    background-color: var(--dark);  
     position: fixed;
     display: flex;
     flex-direction: column;
@@ -107,4 +107,11 @@
     background: rgba(255, 255, 255, 0.04);
     border: 1px solid rgba(255, 255, 255, 0.3);
     border-radius: 8px;
+}
+
+@media screen and  (max-width: 768px){ 
+    .nsidebar {
+        display: none;
+    }
+    
 }
