--- src/components/sidebar/sidebar.css
@@ -1,61 +1,0 @@
-.sidebar-container {
-    margin-top: 3px;
-    background-color: var(--dark);
-    display: flex;
-    color: white;
-    flex-direction: column;
-    margin: 30px;
-}
-.sidebar-title {
-    display: flex;
-    justify-content: space-between;
-    margin-top: 32px;
-    margin-bottom: 50px;
-}
-
-.sidebar-title > span {
-    font-weight: bold;
-    margin-left: 18px;
-    width: 100%;
-}
-/* .sidebar-list > li:hover {
-    font-weight: bold;
-    font-size: 17px;
-} */
-
-.sidebar-list > li {
-    display: flex; 
-    margin-bottom: 35px;
-    align-items: center;
-
-}
-.sidebar-list > li > a {
-    color: white;
-    display: flex;
-    flex-direction: row;
-}
-
-.sidebar-list > li > a> img {
-    margin-right: 18px;
-}
-
-.sidebar-list > li > a:hover {
-    font-weight: bold;
-}
-
-.sidebar-container > button {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    width: 100%;
-    padding: 8px;
-    background: rgba(255, 255, 255, 0.04);
-    border: 1px solid rgba(255, 255, 255, 0.3);
-    border-radius: 8px;
-    color: white;
-    margin-top: 46px;
-    margin-bottom: 64px;
-}
-.sidebar-container > button > span {
-    margin-left: 8px;
-}
