--- src/pages/thedashboard.css
@@ -1,82 +1,0 @@
-.dashboard {
-    /* width: 100%; */
-    height: 100vh;
-    padding: 30px 70px;
-    margin-left: 300px;
-    background: #D7D7D7;
-}
-.mobile-dashboard {
-    display: flex;
-}
-.desktop-dashboard {
-    display: none;
-}
-
-@media (min-width: 760px) {
-    .mobile-dashboard {
-        display: none;
-    }
-    .desktop-dashboard {
-        display: block;
-    }
-    header {
-        display: flex;
-        flex-direction: row;
-        justify-content: space-between;
-        align-items: center;
-    }
-    header button {
-        width: 193px;
-        height: 40px;
-        color: #ffffff;
-        background: #20A352;
-        border-radius: 10px;
-        outline: none;
-        border: none;
-    }
-    header input {
-        width: 383px;
-        height: 40px;
-        background: #FFFFFF;
-        border-radius: 12px;
-        padding: 10px;
-        outline: none;
-        border: none;
-    }
-    .button-input {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        column-gap: 20px;
-    }
-    .desktop-dashboard-overview {
-        margin-top: 100px;
-    }
-    .desktop-dashboard-overview > p {
-        font-weight: 500;
-        font-size: 18px;
-        line-height: 22px;
-        color: #111111;
-        margin-bottom: 10px;
-    }
-    .desktop-dashboard-overview > .overview-tags {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-    }
-    .overview-tags > .overview-card {
-        display: flex;
-        justify-content: space-evenly;
-        align-items: center;
-        width: 245px;
-        height: 100px;
-        background: #FFFFFF;
-        border-radius: 12px;
-    }
-    .overview-card .overview-span {
-        display: flex;
-        flex-direction: column;
-        justify-content: center;
-        align-items: flex-start;
-    }
-}
