+++ src/routes.js
@@ -1,5 +1,5 @@
 import {BrowserRouter, Routes, Route} from "react-router-dom";
-import Sidebar from "./components/sidebar/Sidebar"
+// import Sidebar from "./components/sidebar/Sidebar"
 import NewSidebar from "./components/newsidebar/NewSidebar";
 import Layout from "./pages/Layout"
 import Home from "./pages/Home"
@@ -8,7 +8,7 @@ import Personal from './pages/Personal'
 import Signin from "./pages/Signin"
 import Signup from "./pages/Signup"
 import Errorpage from "./pages/Errorpage"
-import DashBoard from "./pages/dashboard/Dashboard"
+// import DashBoard from "./pages/dashboard/Dashboard"
 import Settings from "./pages/dashboard/settings/Settings"
 import TheDashboard from "./pages/TheDashboard"
 import NewSettings from "./pages/NewSettings";
@@ -23,7 +23,7 @@ const DashboardRoutes = () => {
         <div className="app-routes">
             <NewSidebar isActive={isActive} />
             <Routes>
-                <Route path="the-dashboard" element={<TheDashboard />} />
+                <Route path="dashboard" element={<TheDashboard />} />
                 <Route path="new-settings" element={<NewSettings />} />
                 <Route path="account" element={<Account />} />
                 <Route path="payment" element={<Payment />} />
@@ -45,7 +45,6 @@ const AppRoutes = () => {
                 <Route path="Signin" element={<Signin />} />
                 <Route path="Signup" element={<Signup />} />
                 {/* <Route path="*" element={<Errorpage />} /> */}
-                <Route path="dashboard" element={<DashBoard/>} />
                 <Route path="settings" element={<Settings/>} />
             </Route>
           </Routes>
